Reconcile VPC deletion
