+++ pkg/aws/resource.go
@@ -4,18 +4,6 @@ import (
 	"sigs.k8s.io/cluster-api/util/conditions"
 )
 
-type CloudResourceRequest[TResourceSpec any] struct {
-	RoleARN string
-	Region  string
-	Spec    TResourceSpec
-}
-
-type CloudResourcesRequest[TResourceSpec any] struct {
-	RoleARN string
-	Region  string
-	Specs   []TResourceSpec
-}
-
 type ReconcileRequest[TResourceSpec any] struct {
 	CloudResourceRequest[TResourceSpec]
 
@@ -25,26 +13,12 @@ type ReconcileRequest[TResourceSpec any] struct {
 	ClusterName string
 }
 
-// ReconcileDeleteRequest is a generic request object for deleting a resource with ID.
-type ReconcileDeleteRequest struct {
-	CloudResourceRequest[DeletedCloudResource]
-
-	// Resource that is being reconciled.
-	Resource conditions.Setter
-
-	ClusterName string
+type CloudResourceRequest[TResourceSpec any] struct {
+	RoleARN string
+	Region  string
+	Spec    TResourceSpec
 }
 
-// ReconcileDeleteAllRequest is a generic request object for deleting resources with IDs.
-type ReconcileDeleteAllRequest struct {
-	CloudResourcesRequest[DeletedCloudResource]
-
-	// Resource that is being reconciled.
-	Resource conditions.Setter
-
-	ClusterName string
-}
-
-type DeletedCloudResource struct {
+type DeletedCloudResourceSpec struct {
 	Id string
 }
